requirements.txt
pandas==1.0.1
python-dotenv==0.15.0
snowflake==0.0.3
snowflake-connector-python==2.2.10
SQLAlchemy==1.4.0
snowflake-sqlalchemy==1.2.4
pytest==5.3.5
